[Decomp][Player] Cohesion split: extract death methods to PlayerDeath.cpp
